requirements.txt
selenium
webdriver_manager
beautifulsoup4
python-dotenv
lxml
